+++ app/education/page.jsx
@@ -1,0 +1,33 @@
+"use client";
+
+import React from "react";
+import { useEffect, useState } from "react";
+
+const page = () => {
+  const [cursorPosition, setCursorPosition] = useState({ x: 0, y: 0 });
+
+  useEffect(() => {
+    const handleMouseMove = (event) => {
+      setCursorPosition({ x: event.clientX, y: event.clientY });
+    };
+
+    window.addEventListener("mousemove", handleMouseMove);
+    return () => window.removeEventListener("mousemove", handleMouseMove);
+  }, []);
+
+  return (
+    <>
+      <div>Hello, Education.</div>
+
+      <div
+        className="cursor"
+        style={{
+          left: `${cursorPosition.x}px`,
+          top: `${cursorPosition.y}px`,
+        }}
+      />
+    </>
+  );
+};
+
+export default page;
