+++ components/cards/SushiScroll.jsx
@@ -1,18 +1,18 @@
 import React from "react";
-import "../../styles/osncard.css";
+import "@/styles/sushiscroll.css";
 
-const OSNCard = () => {
+const SushiScroll = () => {
   return (
     <div
-      className="openSpotNotifierCard"
+      className="sushiScrollCard"
       onMouseEnter={() => document.body.classList.add("no-cursor")}
       onMouseLeave={() => document.body.classList.remove("no-cursor")}
     >
       <div className="content">
-        <p className="openSpotNotifierText">Open Spot Notifier</p>
+        <p className="sushiScrollText">SUSHI SCROLL</p>
       </div>
     </div>
   );
 };
 
-export default OSNCard;
+export default SushiScroll;
